requirements.txt
tqdm~=4.67.1
scikit-learn~=1.7.0rc1
matplotlib~=3.10.3
pandas~=2.2.3
requests~=2.32.3
aiohttp~=3.11.18
async-timeout~=5.0.1
pillow~=11.2.1
rembg~=2.0.67
onnxruntime~=1.22.1